chore(schema): twitch_streamers auf Identitätstabelle reduziert + is_partner BOOLEAN-Fix
- twitch_streamers: is_monitored_only, discord_user_id, archived_at entfernt
- Neue Tabelle twitch_exclusions (kind: opt_out | banned)
- Datenmigration: fr4gm1nt/snaqeu → opt_out, skifahrertv → banned
- monitored-only-Logik: NOT EXISTS (twitch_partners) statt Flag
- discord_user_id: alle Reads auf twitch_streamer_identities umgestellt
- archived_at: twitch_partners.admin_archived_at (bereits vorhanden)
- is_partner BOOLEAN-Fix in Stats, Market-Queries, Test-Fixtures
- Migration 20260618000000 + Baseline-Update

+++ rust/crates/tb-monitoring/src/stats.rs
@@ -1,7 +1,6 @@
 //! Time-Series-Logging in `twitch_stats_tracked` / `twitch_stats_category`
 //! (PK-lose Insert-only-Tabellen, Schema-Vertrag — kein Dedup, bewusst wie
-//! Python). `ts_utc` ist timestamptz, `is_partner` ist INTEGER (0/1, SQLite-Erbe
-//! — Bind als `i32`, nicht bool).
+//! Python). `ts_utc` ist timestamptz, `is_partner` ist BOOLEAN (Prod-Schema).
 
 use chrono::{DateTime, Utc};
 use sqlx::PgPool;
@@ -68,7 +67,7 @@ impl StatsStore {
                 .bind(ts)
                 .bind(&row.streamer)
                 .bind(row.viewer_count)
-                .bind(i32::from(row.is_partner))
+                .bind(row.is_partner)
                 .bind(&row.game_name)
                 .bind(&row.stream_title)
                 .bind(&row.tags)
